refactor data & config logic into subfolders
